--- src/jest.d.ts
@@ -1,4 +1,0 @@
-declare namespace jest {
-    const isMockProp: IsMockProp;
-    const spyOnProp: SpyOnProp<Spyable>;
-}
