+++ src/utils/logging.ts
@@ -1,0 +1,5 @@
+export const log = (...args: unknown[]): void => log.default(...args);
+// eslint-disable-next-line no-console
+log.default = log.warn = (...args: unknown[]): void => console.warn(...args);
+
+export default log;
